+++ sdl_exp/Axis.h
@@ -2,12 +2,12 @@
 class Axis
 {
 public:
-	Axis(double length = 1.0);
+    Axis(float length = 1.0);
 	~Axis();
 
 	void render();
 
 private:
-	double length;
+    float length;
 };
